+++ js/data.js
@@ -1,0 +1,41 @@
+const DESCRIPTIONS = [
+  'Просто бомба',
+  'Пляж, коктейль, солнце',
+  'Красота природы',
+  'Счастье не за горами',
+  'Момент наслаждения',
+  'Кааааааайф',
+  'Боже, как же это красиво',
+  'Берег и облака',
+  'Одиночество...',
+  'Вот это встреча'
+  ];
+
+const NAMES = [
+  'Виктор',
+  'Алена',
+  'Елена',
+  'Света',
+  'Катерина',
+  'Максим',
+  'Коля'
+  ];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ];
+
+const COUNT_PHOTO = 25;
+const COMMENTS_MAX = 30;
+
+  const Likes = {
+  MIN:15,
+  MAX:200
+  };
+
+export {DESCRIPTIONS, MESSAGES, NAMES, COUNT_PHOTO, COMMENTS_MAX, Likes}
